requirements.txt
PyQt6
pyserial
python-escpos
pyqtgraph
requests
pywin32
qrcode[pil]
pytz
Flask
supabase
yoyo-migrations
mss
opencv-python
sounddevice
scipy
